requirements.txt
cohere
python-dotenv
langchain
langchain_cohere
flask
flask-cors
pymupdf
pymupdf4llm
numpy